move header into components
